Add date in record teaser
